py3 fix
