+++ cockroachdb/tests/test_check.py
@@ -3,10 +3,13 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
 
+from datadog_checks.base.constants import ServiceCheck
 from datadog_checks.cockroachdb import CockroachdbCheck
 from datadog_checks.dev.testing import requires_py3
+from datadog_checks.dev.utils import get_metadata_metrics
 
 from .common import COCKROACHDB_VERSION, assert_metrics
+from .utils import get_fixture_path
 
 pytestmark = [pytest.mark.integration, pytest.mark.usefixtures('dd_environment'), requires_py3]
 
@@ -16,6 +19,20 @@ def test_metrics(aggregator, instance, dd_run_check):
     dd_run_check(check)
 
     assert_metrics(aggregator)
+
+
+def test_security_metrics(aggregator, instance, dd_run_check, mock_http_response):
+
+    get_metadata_metrics()
+    mock_http_response(file_path=get_fixture_path('security_metrics.txt'))
+
+    check = CockroachdbCheck('cockroachdb', {}, [instance])
+    dd_run_check(check)
+
+    assert_metrics(aggregator)
+
+    aggregator.assert_service_check('cockroachdb.openmetrics.health', ServiceCheck.OK)
+    aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
 def test_version_metadata(aggregator, instance, datadog_agent, dd_run_check):
